[Bug] Corrected data block size calculation (#370)
Made changes to correct the data block size calculation based on which decision is taken to created a new file for a particular task

+++ processing/src/main/java/org/carbondata/processing/store/writer/AbstractFactDataWriter.java
@@ -226,7 +226,7 @@ public abstract class AbstractFactDataWriter<T> implements CarbonFactDataWriter<
    *
    * @throws CarbonDataWriterException if any problem
    */
-  protected void updateBlockletFileChannel(int blockletDataSize) throws CarbonDataWriterException {
+  protected void updateBlockletFileChannel(long blockletDataSize) throws CarbonDataWriterException {
     // get the current file size exceeding the file size threshold
     if ((currentFileSize + blockletDataSize) >= dataBlockSize && currentFileSize != 0) {
       // set the current file size to zero
